introduce FeatureEntryBase
