+++ lib/rggen/core.rb
@@ -102,6 +102,7 @@ require_relative 'core/builder/component_registry'
 require_relative 'core/builder/loader_registry'
 require_relative 'core/builder/input_component_registry'
 require_relative 'core/builder/output_component_registry'
+require_relative 'core/builder/feature_entry_base'
 require_relative 'core/builder/simple_feature_entry'
 require_relative 'core/builder/list_feature_entry'
 require_relative 'core/builder/feature_registry'
